% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlignment.R
\name{alignBatch}
\alias{alignBatch}
\title{alignBatch}
\usage{
alignBatch(object, combine.mz.tol = 25, combine.rt.tol = 30)
}
\arguments{
\item{object}{A metflowClass object.}

\item{combine.mz.tol}{m/z tolerance.}

\item{combine.rt.tol}{RT tolerance.}
}
\value{
A new metflowClass object.
}
\description{
Align different batch datasets.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
