% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{calculate_rsd}
\alias{calculate_rsd}
\title{calculate_rsd}
\usage{
calculate_rsd(
  object,
  slot = c("Subject", "QC", "QC.DL", "Blank", "Tags", "peak.table", "sample.info")
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{Class of data.}
}
\value{
A data frame with RSD.
}
\description{
Calculate RSD of peaks.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
