% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calFC}
\alias{calFC}
\title{calFC}
\usage{
calFC(object, control.group, case.group, type = c("median", "mean"))
}
\arguments{
\item{object}{metflowClass object}

\item{control.group}{Control group name.}

\item{case.group}{Case group name.}

\item{type}{type}
}
\value{
Fold change.
}
\description{
Calculate fold change for metflowClass.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
