% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{calRSD}
\alias{calRSD}
\title{calRSD}
\usage{
calRSD(object, slot = c("Subject", "QC"))
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{Class of data.}
}
\value{
A data frame with RSD.
}
\description{
Calculate RSD of peaks.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
