% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{chromatogramPlot}
\alias{chromatogramPlot}
\title{chromatogramPlot}
\usage{
chromatogramPlot(object, title.size = 15, lab.size = 15,
  axis.text.size = 15, alpha = 0.5, title = "")
}
\arguments{
\item{object}{Object for tic.plot or bpc.plot.}

\item{title.size}{Font size of title..}

\item{lab.size}{Font size of lab title.}

\item{axis.text.size}{Font size of axis text.}

\item{alpha}{alpha.}

\item{title}{Title of the plot..}
}
\value{
A ggplot2 object.
}
\description{
Draw TIC or BPC.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
