% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{create_metflow_object}
\alias{create_metflow_object}
\title{create_metflow_object}
\usage{
create_metflow_object(ms1.data, sample.information, path = ".")
}
\arguments{
\item{ms1.data}{MS1 peak table name.}

\item{sample.information}{Sample information name.}

\item{path}{Work directory.}
}
\value{
A metflowClass object.
}
\description{
Create metflowClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
