% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalTranslater.R
\name{databaseName}
\alias{databaseName}
\title{databaseName}
\usage{
databaseName(from.to = c("from", "to"))
}
\arguments{
\item{from.to}{From or to.}
}
\value{
A vector..
}
\description{
Whate databases are supported.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
