% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4data_processing.R
\name{extractEIC}
\alias{extractEIC}
\title{extractEIC}
\usage{
extractEIC(
  object,
  mz.range,
  rt.range,
  title.size = 15,
  lab.size = 15,
  axis.text.size = 15,
  alpha = 0.5,
  title = ""
)
}
\arguments{
\item{object}{object}

\item{mz.range}{mz.range}

\item{rt.range}{rt.range}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size}

\item{alpha}{alpha}

\item{title}{title}
}
\value{
result
}
\description{
extractEIC
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
