% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFilter.R
\name{filterPeak2}
\alias{filterPeak2}
\title{filterPeak}
\usage{
filterPeak2(
  object,
  min.fraction = 0.5,
  type = c("any", "all"),
  min.subject.qc.ratio = 2,
  dl.qc.r2.cutoff = 0.7,
  according.to = c("class", "group"),
  which.group
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{min.fraction}{Peaks minimum fraction in samples.}

\item{type}{Any ("any") or all ("all") groups should be meet this min.fraction.}

\item{min.subject.qc.ratio}{Peak intensity ratio in subject and blank samples.}

\item{dl.qc.r2.cutoff}{R2 cutoff for dilution QC.}

\item{according.to}{Which information you want to use in sample information.}

\item{which.group}{What groups you want to use.}
}
\value{
A new metflowClass object.
}
\description{
Filter peaks.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
