% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{get_data}
\alias{get_data}
\title{Get data from metflowClass object.}
\usage{
get_data(
  object,
  slot = c("Subject", "QC", "QC.DL", "Blank", "Tags", "peak.table", "sample.info")
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{Class of data.}
}
\value{
A data frame.
}
\description{
Get data from metflowClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
