% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classyfire.R
\name{get_metclass}
\alias{get_metclass}
\title{get_metclass}
\usage{
get_metclass(inchkey = "QZDWODWEESGPLC-UHFFFAOYSA-N",
  server = "http://classyfire.wishartlab.com/entities/")
}
\arguments{
\item{inchkey}{The inchkey ID of a metabolite.}

\item{server}{server.}
}
\value{
A classyfire class object.
}
\description{
Get the class information of a metabolite using classyfire.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
