% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{get_mv_plot_samples}
\alias{get_mv_plot_samples}
\title{get_mv_plot_samples}
\usage{
get_mv_plot_samples(object, interactive = FALSE)
}
\arguments{
\item{object}{A metflowClass object.}

\item{interactive}{interactive or not.}
}
\value{
A ggplot2 object.
}
\description{
get MV plot of subject samples.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
