% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(object)
}
\arguments{
\item{object}{A metflowClass object.}
}
\value{
A data frame of parameters.
}
\description{
Get parameters from a metflowClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
