% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{getMVplot4sample}
\alias{getMVplot4sample}
\title{getMVplot4sample}
\usage{
getMVplot4sample(object)
}
\arguments{
\item{object}{A metflowClass object.}
}
\value{
A ggplot2 object.
}
\description{
get MV plot of subject samples.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
