% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCreation.R
\name{getParams}
\alias{getParams}
\title{getParams}
\usage{
getParams(object)
}
\arguments{
\item{object}{A metflowClass object.}
}
\value{
A data frame of parameters.
}
\description{
Get parameters from a metflowClass object.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
