% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{metflow2_logo}
\alias{metflow2_logo}
\title{metflow2_logo}
\usage{
metflow2_logo(colour = c("yellow", "white", "green", "red"))
}
\arguments{
\item{colour}{Logo colour.}
}
\value{
Ascii art logo.
}
\description{
Show logo of metflow2.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
