% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{metflow2_packages}
\alias{metflow2_packages}
\title{List all packages in the metflow2}
\usage{
metflow2_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metflow2 in the list?}
}
\description{
List all packages in the metflow2
}
\examples{
metflow2_packages()
}
