% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataNormalization.R
\name{normalizeData}
\alias{normalizeData}
\title{normalizeData}
\usage{
normalizeData(object, method = c("svr", "total", "median", "mean",
  "pqn"), keep.scale = TRUE)
}
\arguments{
\item{object}{A metflowClass object.}

\item{method}{Normalization method, mean, median, total svr or loess,
default is svr. Please see the details.}

\item{keep.scale}{Remain scale or not. Default is TRUE.}
}
\value{
A new metflowClass object.
}
\description{
Data normalization.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
