% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4data_processing.R
\name{outputFeatureEIC}
\alias{outputFeatureEIC}
\title{outputFeatureEIC}
\usage{
outputFeatureEIC(
  object,
  feature.index = 1,
  polygon = TRUE,
  save.plot = TRUE,
  path = ".",
  file.type = c("pdf", "png"),
  plot.name = "peak_plot",
  width = 7,
  height = 7,
  interactive.plot = FALSE,
  return.plot = TRUE,
  message = TRUE
)
}
\arguments{
\item{object}{xdata class object from dataProcessing.}

\item{feature.index}{Which feature to output. Index.}

\item{polygon}{Add polygon or not.}

\item{save.plot}{Save plot or not.}

\item{path}{Work directory.}

\item{file.type}{pdf or png.}

\item{plot.name}{The name of plot.}

\item{width}{Width.}

\item{height}{Height.}

\item{interactive.plot}{Interactive or not.}

\item{return.plot}{Return plot or not.}

\item{message}{Show message or not.}
}
\value{
ggplot object.
}
\description{
Output EICs of all peaks in peak table.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
