% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4dataProcessing.R
\name{plot_chromatogram}
\alias{plot_chromatogram}
\title{plot_chromatogram}
\usage{
plot_chromatogram(
  object,
  title.size = 13,
  lab.size = 13,
  axis.text.size = 12,
  alpha = 0.5,
  title = "",
  interactive = FALSE,
  group.for.figure = "QC"
)
}
\arguments{
\item{object}{Object for tic.plot or bpc.plot.}

\item{title.size}{Font size of title..}

\item{lab.size}{Font size of lab title.}

\item{axis.text.size}{Font size of axis text.}

\item{alpha}{alpha.}

\item{title}{Title of the plot..}

\item{interactive}{interactive.}

\item{group.for.figure}{What groups to show EIC.}
}
\value{
A ggplot2 object.
}
\description{
Draw TIC or BPC.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
