% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4data_processing.R
\name{plot_tic}
\alias{plot_tic}
\title{plot_tic}
\usage{
plot_tic(file, path = ".", type = c("tic", "bpc"), threads = 4, legend = FALSE)
}
\arguments{
\item{file}{name of mzXML data.}

\item{path}{work directory.}

\item{type}{tic or bpc.}

\item{threads}{thread number}

\item{legend}{legend or not.}
}
\value{
A object for plot_chromatogram() function.
}
\description{
Extract TIC or BPC from in a RT range from mzXML format data.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
