% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4data_processing.R
\name{plotAdjustedRT}
\alias{plotAdjustedRT}
\title{plotAdjustedRT}
\usage{
plotAdjustedRT(object, title.size = 15, lab.size = 15, axis.text.size = 15)
}
\arguments{
\item{object}{object}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size}
}
\value{
A ggplot2 object.
}
\description{
plotAdjustedRT
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
