% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicated_samples.R
\name{remove_duplicated_name}
\alias{remove_duplicated_name}
\title{Remove the duplicated samples of peak tables from Thermo QE instrument}
\usage{
remove_duplicated_name(peak_table)
}
\arguments{
\item{peak_table}{A peak_table from XCMS or other software.}
}
\value{
A new peak table without duplicated samples.
}
\description{
Remove the duplicated samples of peak tables from Thermo QE instrument.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
