% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicated_samples.R
\name{show_duplicated_samples}
\alias{show_duplicated_samples}
\title{Show the duplicated samples of peak tables from Thermo QE instrument}
\usage{
show_duplicated_samples(peak_table)
}
\arguments{
\item{peak_table}{A peak_table from XCMS or other software.}
}
\value{
A list of duplicated samples.
}
\description{
Show the duplicated samples of peak tables from Thermo QE instrument.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
