% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{showPeak}
\alias{showPeak}
\title{showPeak}
\usage{
showPeak(object, peak.index = 1, title.size = 15, lab.size = 15,
  axis.text.size = 15, alpha = 0.5, title = "", interactive = TRUE)
}
\arguments{
\item{object}{Object from extractPeaks.}

\item{peak.index}{Which peak to show. Index.}

\item{title.size}{Title size.}

\item{lab.size}{Lab titile size.}

\item{axis.text.size}{Text size of axis.}

\item{alpha}{alpha.}

\item{title}{Title of the plot.}

\item{interactive}{Interactive or not.}
}
\value{
Result contains EIC of peaks.
}
\description{
Show the peaks from result from extractPeaks function.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
