% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{showPeakPlot}
\alias{showPeakPlot}
\title{showPeakPlot}
\usage{
showPeakPlot(object1, object2, peak.index = 1)
}
\arguments{
\item{object1}{metflowClass object 1.}

\item{object2}{metflowClass object 2.}

\item{peak.index}{Peak index}
}
\value{
Peak plot (ggplot2 object).
}
\description{
Show peak plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
