% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sxtScale}
\alias{sxtScale}
\title{sxtScale}
\usage{
sxtScale(df, method = c("no", "auto", "pareto", "center"))
}
\arguments{
\item{df}{df}

\item{method}{method}
}
\value{
result
}
\description{
sxtScale
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
