% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalTranslater.R
\name{transID}
\alias{transID}
\title{transID}
\usage{
transID(query = "C00001", from = "KEGG", to = "PubChem SID",
  top = 1, server = "http://cts.fiehnlab.ucdavis.edu/service/convert")
}
\arguments{
\item{query}{The ID of metabolite you want to translate.}

\item{from}{The databases of metabolites. Supported database can be shown using databaseName("from").}

\item{to}{The databases of metabolites. Supported database can be shown using databaseName("to").}

\item{top}{How many results should be returned?}

\item{server}{server.}
}
\value{
A data frame.
}
\description{
Translate metabolite ID.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
