% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{VolcanoPlot}
\usage{
volcanoPlot(
  p.value,
  fc,
  p.cutoff = 0.05,
  fc.cutoff = 1.3,
  xlab = "Log2 (Fold change)",
  ylab = "-Log10 (P-value)"
)
}
\arguments{
\item{p.value}{p.value.}

\item{fc}{fc.}

\item{p.cutoff}{p.cutoff.}

\item{fc.cutoff}{fc.cutoff.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}
}
\value{
volcano plot.
}
\description{
Draw volcano plot.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
