% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calP}
\alias{calP}
\title{UnivariateTest}
\usage{
calP(object, control.group, case.group, test.method = c("t", "wilcox"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH",
  "BY", "fdr", "none"), log.scale = FALSE, ...)
}
\arguments{
\item{test.method}{Which test you want to use? "t" means stutent t test
and "wilcox" mean wilcoxon test.}

\item{adjust.method}{p value correction method. See p.adjust function.}

\item{log.scale}{Data transformation method, defaulst is FALSE.}

\item{MetFlowData}{MetFlowData.}

\item{class}{Class used to do test.}
}
\value{
MetFlowData which has been added p and AUC information in tags.
}
\description{
Calculate p value and AUC for each feature.
}
\seealso{
The details of univariate test can
be found in \code{\link[stats]{t.test}},
\code{\link[stats]{p.adjust}} and \code{\link[stats]{wilcox.test}}.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
