% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calP}
\alias{calP}
\title{calP}
\usage{
calP(object, control.group, case.group, test.method = c("t", "wilcox"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH",
  "BY", "fdr", "none"), log.scale = FALSE, ...)
}
\arguments{
\item{object}{object}

\item{control.group}{control.group.}

\item{case.group}{case.group.}

\item{test.method}{test.method.}

\item{adjust.method}{adjust.method.}

\item{log.scale}{log.scale.}

\item{...}{other arguments.}
}
\value{
p values.
}
\description{
Calculate p value and AUC for each feature.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
