% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalTranslater.R
\name{databaseName}
\alias{databaseName}
\title{databaseName}
\usage{
databaseName(
  server = c("http://cts.fiehnlab.ucdavis.edu/service/convert",
    "https://www.chemspider.com/InChI.asmx")
)
}
\arguments{
\item{server}{server.}
}
\value{
A vector..
}
\description{
Whate databases are supported.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
