% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4dataProcessing.R
\name{extractPeaks}
\alias{extractPeaks}
\title{extractPeaks}
\usage{
extractPeaks(
  path = ".",
  ppm = 15,
  threads = 4,
  is.table = "is.xlsx",
  mz = NULL,
  rt = NULL,
  rt.tolerance = 40
)
}
\arguments{
\item{path}{Work directory.}

\item{ppm}{see xcms.}

\item{threads}{Number of threads.}

\item{is.table}{Peak table. Two columns, column 1 is name of peak, column 2 is m/z of peaks.}

\item{mz}{mz}

\item{rt}{rt}

\item{rt.tolerance}{Rt tolerance.}
}
\value{
Result contains EIC of peaks.
}
\description{
From mzXML data extract peaks according to IS table.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
