% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classyfire.R
\name{get_metclass}
\alias{get_metclass}
\title{get_metclass}
\usage{
get_metclass(inchikey = "QZDWODWEESGPLC-UHFFFAOYSA-N",
  server = "http://classyfire.wishartlab.com/entities/", sleep = 5)
}
\arguments{
\item{inchikey}{The inchikey ID of a metabolite.}

\item{server}{server.}

\item{sleep}{Sleep time for system. Second.}
}
\value{
A classyfire class object.
}
\description{
Get the class information of a metabolite using classyfire.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
