% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(
  object,
  slot = c("Subject", "QC", "QC.DL", "Blank", "Tags"),
  silence.deprecated = TRUE
)
}
\arguments{
\item{object}{A metflowClass object.}

\item{slot}{Class of data.}

\item{silence.deprecated}{Silence deprecated information or not.}
}
\value{
A data frame.
}
\description{
Get data from metflowClass object.
\lifecycle{deprecated}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
