% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metflow2_logo.R, R/tools.R
\name{metflow2_logo}
\alias{metflow2_logo}
\title{metflow2_logo}
\usage{
metflow2_logo(logo.index = 1, colour = c("yellow", "white", "green", "red"))

metflow2_logo(logo.index = 1, colour = c("yellow", "white", "green", "red"))
}
\arguments{
\item{logo.index}{Which logo you want to use.}

\item{colour}{Logo colour.}
}
\value{
Ascii art logo.
}
\description{
metflow2_logo

Show logo of metflow2.
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}

Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
