% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{processData}
\alias{processData}
\title{processData}
\usage{
processData(path = ".", polarity = c("positive", "negative"),
  ppm = 15, peakwidth = c(5, 30), snthresh = 10, mzdiff = -0.001,
  noise = 500, threads = 4, output.tic = TRUE, output.bpc = TRUE,
  output.rt.correction.plot = TRUE, min.fraction = 0.8,
  fill.peaks = FALSE)
}
\arguments{
\item{path}{Work directory.}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{ppm}{see xcms.}

\item{peakwidth}{See xcms.}

\item{snthresh}{See xcms.}

\item{mzdiff}{See xcms.}

\item{noise}{See xcms.}

\item{threads}{Number of threads.}

\item{output.tic}{Output TIC plot or not.}

\item{output.bpc}{Output BPC plot or not.}

\item{output.rt.correction.plot}{Output rt correction plot or not.}

\item{min.fraction}{See xcms.}

\item{fill.peaks}{Fill peaks NA or not.}
}
\value{
Peak table.
}
\description{
Raw MS data processing using xcms.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
