% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{VolcanoPlot}
\usage{
volcanoPlot(p.value, fc, p.cutoff = 0.05, fc.cutoff = 1.3,
  xlab = "Log2 (Fold change)", ylab = "-Log10 (P-value)")
}
\arguments{
\item{p.cutoff}{p value cutoff.}

\item{MetFlowData}{MetFlowData.}

\item{x}{x axis factor.}

\item{y}{y axis factor.}

\item{z}{z axis factor.}

\item{col}{Colour for markers and non-markers.}

\item{foldchange.cutoff}{Fold chagne cutoff.}

\item{vip.cutoff}{VIP value cutoff.}

\item{path}{The directory you want to write results.}
}
\value{
volcano plot.
}
\description{
Draw volcano plot.
}
\examples{
\donttest{
}
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
